+++ weblm/controllers/__init__.py
@@ -1,0 +1,3 @@
+from weblm.controllers.basic.controller import Controller as BasicController
+
+registry = {"basic": BasicController}
